0.1.45 - 2015-10-31 - Improved memory tracking (windows only), fixed contrib warnings, fixed SimpleBuffers, added CX version   - Fixed contrib warnings   - Updated CX version   - Fixed BINReader & BINWriter   - Fixed JSONReader & JSONWriter   - Improved memory tracking (windows only)

+++ Src/Archive/LZ4InputFilter.cpp
@@ -71,7 +71,8 @@ Status LZ4InputFilter::Filter(const void *pInput, Size cbInputSize, Size cbOrigI
 	{
 		return status;
 	}
-	if (0 == (cbResSize = LZ4_decompress_safe((const char *)pInput, (char *)m_buffer.GetMem(), (int)cbInputSize, *pcbOutputSize)))
+	if (0 == (cbResSize = LZ4_decompress_safe((const char *)pInput, (char *)m_buffer.GetMem(), (int)cbInputSize, 
+	                                          (int)*pcbOutputSize)))
 	{
 		return Status_OperationFailed;
 	}
